+++ test/Armada.Test.Database/DatabaseTestRunner.cs
@@ -149,6 +149,7 @@ namespace Armada.Test.Database
             OperationalRoundTripDatabaseTests roundTrips = new OperationalRoundTripDatabaseTests(_Driver, _Settings, _NoCleanup);
             await RunTest("Skill_Create_Read_Update_Reopen_Window_Delete", "Operational", () => roundTrips.VerifySkillsAsync(token), token);
             await RunTest("ProjectProfile_Create_Read_Update_Reopen_Window_Delete", "Operational", () => roundTrips.VerifyProjectProfilesAsync(token), token);
+            await RunTest("Playbook_Create_Read_And_Empty_Text_Reads_As_Null", "Operational", () => roundTrips.VerifyPlaybookEmptyTextAsync(token), token);
             await RunTest("Pipeline_Update_And_Delete_Roll_Back_On_Failure", "Operational", () => TestPipelineWriteAtomicityAsync(token), token);
             await RunTest("RequestHistory_Timestamp_RoundTrip_And_Same_Day_Range", "Operational", () => TestRequestHistorySameDayRangeAsync(token), token);
             await RunTest("CheckRun_Create_Read_Update_Enumerate", "Operational", () => TestCheckRunCrudAsync(token), token);